:recycle: Extract walkToText
